move consul filter functions to new file
Making room for more consul functions by moving these into a more
specifically named file. Tests are also migrated similarly with a couple
more small refactors. Also the general fakeWatcher moved to the module
level test file.
