+++ tfunc/consul_filter.go
